HasApplication and HasContainer traits
