+++ src/Illuminate/Broadcasting/BroadcastManager.php
@@ -11,6 +11,7 @@ use Illuminate\Broadcasting\Broadcasters\NullBroadcaster;
 use Illuminate\Broadcasting\Broadcasters\PusherBroadcaster;
 use Illuminate\Broadcasting\Broadcasters\RedisBroadcaster;
 use Illuminate\Bus\UniqueLock;
+use Illuminate\Container\Concerns\HasApplication;
 use Illuminate\Contracts\Broadcasting\Factory as FactoryContract;
 use Illuminate\Contracts\Broadcasting\ShouldBeUnique;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcastNow;
@@ -26,12 +27,7 @@ use Pusher\Pusher;
  */
 class BroadcastManager implements FactoryContract
 {
-    /**
-     * The application instance.
-     *
-     * @var \Illuminate\Contracts\Container\Container
-     */
-    protected $app;
+    use HasApplication;
 
     /**
      * The array of resolved broadcast drivers.
@@ -484,19 +480,6 @@ class BroadcastManager implements FactoryContract
     public function getApplication()
     {
         return $this->app;
-    }
-
-    /**
-     * Set the application instance used by the manager.
-     *
-     * @param  \Illuminate\Contracts\Foundation\Application  $app
-     * @return $this
-     */
-    public function setApplication($app)
-    {
-        $this->app = $app;
-
-        return $this;
     }
 
     /**
