+++ src/Illuminate/Container/Concerns/HasApplication.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Illuminate\Container\Concerns;
+
+trait HasApplication
+{
+    /**
+     * The application instance.
+     *
+     * @var \Illuminate\Contracts\Foundation\Application
+     */
+    protected $app;
+
+    /**
+     * Set the application instance used by the manager.
+     *
+     * @param  \Illuminate\Contracts\Foundation\Application  $app
+     * @return $this
+     */
+    public function setApplication($app)
+    {
+        $this->app = $app;
+
+        return $this;
+    }
+}
