+++ src/Illuminate/Container/Concerns/HasContainer.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Illuminate\Container\Concerns;
+
+trait HasContainer
+{
+    /**
+     * The container instance.
+     *
+     * @var \Illuminate\Contracts\Container\Container
+     */
+    protected $container;
+
+    /**
+     * Set the container instance used by the gate.
+     *
+     * @param  \Illuminate\Contracts\Container\Container  $container
+     * @return $this
+     */
+    public function setContainer(Container $container)
+    {
+        $this->container = $container;
+
+        return $this;
+    }
+}
